Refactored code to handle sync vs. async calls inside of single struct.
