+++ src/collection.rs
@@ -4,9 +4,14 @@ use super::{
 };
 use serde_json::Value;
 use std::sync::Arc;
+
+#[cfg(feature = "sync")]
+use std::sync::Mutex;
+
+#[cfg(not(feature = "sync"))]
 use tokio::sync::Mutex;
 
-type DocumentCollection = Arc<Mutex<Vec<Value>>>;
+pub type DocumentCollection = Arc<Mutex<Vec<Value>>>;
 
 #[derive(Clone)]
 pub struct Collection {
@@ -20,6 +25,20 @@ impl Collection {
     }
   }
 
+  #[cfg(feature = "sync")]
+  pub fn insert(&self, document: Value) -> Result<(), Error> {
+    if !document.is_object() {
+      return Err(Error::MQError(String::from(
+        "Document must be a JSON object.",
+      )));
+    }
+
+    self.data.lock().unwrap().push(document);
+
+    Ok(())
+  }
+
+  #[cfg(not(feature = "sync"))]
   pub async fn insert(&self, document: Value) -> Result<(), Error> {
     if !document.is_object() {
       return Err(Error::MQError(String::from(
@@ -32,11 +51,23 @@ impl Collection {
     Ok(())
   }
 
+  #[cfg(feature = "sync")]
+  pub fn find(&self, query: Value) -> Result<Documents, Error> {
+    if !query.is_object() {
+      return Err(Error::MQError(String::from("Query must be a JSON object.")));
+    }
+
+    Engine::with_collection(self.data.clone()).find(&query)
+  }
+
+  #[cfg(not(feature = "sync"))]
   pub async fn find(&self, query: Value) -> Result<Documents, Error> {
     if !query.is_object() {
       return Err(Error::MQError(String::from("Query must be a JSON object.")));
     }
 
-    Engine::with_collection(&*self.data.lock().await).find(&query)
+    Engine::with_collection(self.data.clone())
+      .find(&query)
+      .await
   }
 }
