+++ src/memdb.rs
@@ -1,6 +1,11 @@
 use super::{collection::Collection, errors::Error};
 use std::collections::HashMap;
 use std::sync::Arc;
+
+#[cfg(feature = "sync")]
+use std::sync::Mutex;
+
+#[cfg(not(feature = "sync"))]
 use tokio::sync::Mutex;
 
 pub struct MemDb {
@@ -14,6 +19,7 @@ impl MemDb {
     }
   }
 
+  #[cfg(not(feature = "sync"))]
   pub async fn create_collection(&self, name: &str) {
     let new_collection = Collection::new();
     self
@@ -23,6 +29,7 @@ impl MemDb {
       .insert(name.to_string(), new_collection);
   }
 
+  #[cfg(not(feature = "sync"))]
   pub async fn collection(&self, collection_name: &str) -> Result<Collection, Error> {
     match self.collections.lock().await.get(collection_name) {
       Some(c) => Ok(c.clone()),
@@ -30,11 +37,40 @@ impl MemDb {
     }
   }
 
+  #[cfg(not(feature = "sync"))]
   pub async fn delete_collection(&self, name: &str) -> Result<Collection, Error> {
     self
       .collections
       .lock()
       .await
+      .remove(name)
+      .ok_or(Error::MQCollectionNotFound)
+  }
+
+  #[cfg(feature = "sync")]
+  pub fn create_collection(&self, name: &str) {
+    let new_collection = Collection::new();
+    self
+      .collections
+      .lock()
+      .unwrap()
+      .insert(name.to_string(), new_collection);
+  }
+
+  #[cfg(feature = "sync")]
+  pub fn collection(&self, collection_name: &str) -> Result<Collection, Error> {
+    match self.collections.lock().unwrap().get(collection_name) {
+      Some(c) => Ok(c.clone()),
+      None => Err(Error::MQCollectionNotFound),
+    }
+  }
+
+  #[cfg(feature = "sync")]
+  pub fn delete_collection(&self, name: &str) -> Result<Collection, Error> {
+    self
+      .collections
+      .lock()
+      .unwrap()
       .remove(name)
       .ok_or(Error::MQCollectionNotFound)
   }
@@ -44,6 +80,7 @@ impl MemDb {
 mod tests {
   use crate::{errors::Error, memdb::MemDb};
 
+  #[cfg(not(feature = "sync"))]
   #[tokio::test]
   async fn test_create_collection() -> Result<(), Error> {
     let memdb = MemDb::new();
@@ -52,6 +89,7 @@ mod tests {
     Ok(())
   }
 
+  #[cfg(not(feature = "sync"))]
   #[tokio::test]
   async fn test_delete_collection() -> Result<(), Error> {
     let memdb = MemDb::new();
@@ -59,6 +97,29 @@ mod tests {
     let _ = memdb.collection("TestCollection").await?;
     memdb.delete_collection("TestCollection").await?;
     if let Ok(_) = memdb.delete_collection("TestCollection").await {
+      assert_eq!("should not find collection", "found collection");
+    }
+
+    Ok(())
+  }
+
+  #[cfg(feature = "sync")]
+  #[test]
+  fn test_create_collection() -> Result<(), Error> {
+    let memdb = MemDb::new();
+    memdb.create_collection("TestCollection");
+    let _ = memdb.collection("TestCollection")?;
+    Ok(())
+  }
+
+  #[cfg(feature = "sync")]
+  #[test]
+  fn test_delete_collection() -> Result<(), Error> {
+    let memdb = MemDb::new();
+    memdb.create_collection("TestCollection");
+    let _ = memdb.collection("TestCollection")?;
+    memdb.delete_collection("TestCollection")?;
+    if let Ok(_) = memdb.delete_collection("TestCollection") {
       assert_eq!("should not find collection", "found collection");
     }
